Add D3 visualization
For the previous version of this tool, I calculated the average trip
duration for all possible station combinations and stored the results in
a database. In the updated version, every individual trip is stored in
the database, allowing the tool to display a D3 histogram of all the
trip durations between any combination of stations, instead of just the
average duration.